requirements.txt
libclang-python
openai>=1.0.0
anthropic>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
pydantic>=2.0
loguru>=0.7.0
argparse
